Reorganise to minimise dependencies of instrument server
