+++ src/murfey/instrument_server/__init__.py
@@ -5,7 +5,7 @@ import uvicorn
 from rich.logging import RichHandler
 
 import murfey
-from murfey.server import LogFilter
+from murfey.util import LogFilter
 
 logger = logging.getLogger("murfey.instrument_server")
