+++ src/murfey/util/models.py
@@ -347,3 +347,8 @@ class MultigridWatcherSetup(BaseModel):
 
 class CurrentGainRef(BaseModel):
     path: str
+
+
+class Token(BaseModel):
+    access_token: str
+    token_type: str
